carlitos/mcp-servers/image-gen/requirements.txt
openai==1.77.0
fastapi==0.115.12
uvicorn==0.34.2
sse-starlette==2.3.4
pydantic==2.11.4
pillow==11.2.1
httpx==0.28.1
mcp==1.7.1